+++ src/workers/backgroundTaskWorker.ts
@@ -1,11 +1,11 @@
-import { verifyIpfsData } from '@src/utils/ipfs';
+import { verifyIpfsData } from '@src/libs/ipfs';
 
 export interface CommentPayload {
   uri: string;
   pendingCommentId: string;
 }
 
- 
+
 self.onmessage = async (event: MessageEvent<{ type: string; payload: CommentPayload }>) => {
   const { type, payload } = event.data;
 
@@ -17,11 +17,11 @@ self.onmessage = async (event: MessageEvent<{ type: string; payload: CommentPayl
       await verifyIpfsData(uri);
 
       // Send a success message to main thread
-       
+
       self.postMessage({ success: true, pendingCommentId });
     } catch (error) {
       // Send a error message to main thread
-       
+
       self.postMessage({ success: false, error: (error as Error).message, pendingCommentId });
     }
   }
